multi_accounts/pyproject.toml
[tool.poetry]
authors = ["Andre Sionek <andre@sionek.com>"]
name = "best_practices"
description = "Multi Accounts"
keywords = ['cdk', 'iac', 'infrastructure', 'devops', 'python', 'backend', "platform"] # noqa
readme = "README.md"
homepage = "https://sionek.com/"
version = "0.0.0"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.99.0"
